rename transport to 'protocol'
This seems a little less confusing in server code.
